fix and UIView+Border
